# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_documents.documentListItem:
    properties:
      analyzed_at:
        example: "2026-06-10T12:03:00Z"
        type: string
      file_name:
        example: contract.pdf
        type: string
      id:
        example: doc_001
        format: uuid
        type: string
      risks:
        example: 3
        type: integer
      summary:
        example: Analyzed legal document
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@mengu.ai
    name: API Support
  description: AI-assisted email automation platform. Ingests emails, analyzes them
    with LLM to extract intent and actions, then executes actions through deterministic
    handlers.
  license:
    name: Proprietary
  termsOfService: https://mengu.ai/terms
  title: Mengu AI API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user via email and password. Returns JWT access and
        refresh tokens.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              refresh_token:
                type: string
              token_type:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      summary: Authenticate user
      tags:
      - Authentication
  /auth/oauth/google:
    post:
      consumes:
      - application/json
      description: Authenticate or register via Google OAuth2 authorization code.
      parameters:
      - description: Google OAuth authorization code
        in: body
        name: request
        required: true
        schema:
          properties:
            code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              refresh_token:
                type: string
              token_type:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      summary: Google OAuth login
      tags:
      - Authentication
  /auth/oauth/microsoft:
    post:
      consumes:
      - application/json
      description: Authenticate or register via Microsoft OAuth2 authorization code.
      parameters:
      - description: Microsoft OAuth authorization code
        in: body
        name: request
        required: true
        schema:
          properties:
            code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              refresh_token:
                type: string
              token_type:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      summary: Microsoft OAuth login
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access/refresh token pair.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_in:
                type: integer
              refresh_token:
                type: string
              token_type:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      summary: Refresh access token
      tags:
      - Authentication
...
  /events/{id}/documents:
    get:
      consumes:
      - application/json
      description: List AI-analyzed documents associated with a specific event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_documents.documentListItem'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: List analyzed documents for event
      tags:
      - Documents
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  WebhookSecret:
    description: Webhook secret key for authenticating email webhook requests.
    in: header
    name: X-Webhook-Secret
    type: apiKey
swagger: "2.0"